доработал стили корзины

+++ cart/models.py
@@ -6,7 +6,7 @@ from products.models import Product
 class CartItem(models.Model):
     user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
     product = models.ForeignKey(Product, on_delete=models.CASCADE)
-    quantity = models.PositiveBigIntegerField(default=1)
+    quantity = models.PositiveIntegerField(default=1)
     added_at = models.DateTimeField(auto_now_add=True)
 
     class Meta:
@@ -18,5 +18,3 @@ class CartItem(models.Model):
     @property
     def total_price(self):
         return self.product.price * self.quantity
-
-
